[REFACTOR] move clis to clis subpackage
