+++ src/biocatalyzer/clis/cli_bioreactor.py
@@ -3,8 +3,7 @@ import os
 
 import click
 
-from biocatalyzer.bioreactor import BioReactor
-from biocatalyzer.matcher import MSDataMatcher
+from biocatalyzer import BioReactor
 
 DATA_FILES = os.path.dirname(__file__)
 
@@ -56,32 +55,6 @@ DATA_FILES = os.path.dirname(__file__)
               show_default=True,
               help="The minimum atom count of a molecule (molecules with less atoms are removed from the products).",
               )
-@click.option("--match_ms_data",
-              "match_ms_data",
-              type=bool,
-              default=False,
-              show_default=True,
-              help="Whether to match the generated products with MS data.")
-@click.option("--ms_data_path",
-              "ms_data_path",
-              type=click.File('r'),
-              default=None,
-              show_default=True,
-              help="The path to the file containing the MS data to use.")
-@click.option("--mode",
-              "mode",
-              type=click.Choice(['mass', 'mass_diff']),
-              default='mass',
-              show_default=True,
-              help="The mode to use for the MS data matching (mass for ExactMass matching or mass_dif for ExactMass "
-                   "differences matching).")
-@click.option("--tolerance",
-              "tolerance",
-              type=float,
-              default=0.02,
-              show_default=True,
-              help="The mass tolerance to use when matching MS data.",
-              )
 @click.option("--n_jobs",
               "n_jobs",
               type=int,
@@ -89,20 +62,16 @@ DATA_FILES = os.path.dirname(__file__)
               show_default=True,
               help="The number of jobs to run in parallel.",
               )
-def main(compounds,
-         output_path,
-         neutralize,
-         reaction_rules,
-         organisms,
-         patterns_to_remove,
-         molecules_to_remove,
-         min_atom_count,
-         match_ms_data,
-         ms_data_path,
-         mode,
-         tolerance,
-         n_jobs):
-    """Run the BioCatalyzer and the MSDataMatcher (optional).
+def bioreactor_cli(compounds,
+                   output_path,
+                   neutralize,
+                   reaction_rules,
+                   organisms,
+                   patterns_to_remove,
+                   molecules_to_remove,
+                   min_atom_count,
+                   n_jobs):
+    """Run the BioCatalyzer.
 
     Mandatory arguments:
 
@@ -112,7 +81,7 @@ def main(compounds,
     """
     if reaction_rules is None:
         reaction_rules = os.path.join(
-                DATA_FILES, 'data/reactionrules/all_reaction_rules_forward_no_smarts_duplicates_sample.tsv')
+            DATA_FILES, '../data/reactionrules/all_reaction_rules_forward_no_smarts_duplicates_sample.tsv')
     br = BioReactor(compounds_path=compounds,
                     output_path=output_path,
                     reaction_rules_path=reaction_rules,
@@ -124,18 +93,6 @@ def main(compounds,
                     n_jobs=n_jobs)
     logging.basicConfig(filename=f'{output_path}logging.log', level=logging.DEBUG)
     br.react()
-
-    if match_ms_data:
-        if not ms_data_path:
-            raise ValueError("The path to the MS data file is required when matching MS data.")
-
-        ms = MSDataMatcher(ms_data_path=ms_data_path,
-                           compounds_to_match=f"{output_path}/new_compounds.tsv",
-                           output_path=output_path,
-                           mode=mode,
-                           tolerance=tolerance)
-
-        ms.generate_ms_results()
 
 
 if __name__ == "__main__":
